viz/pyproject.toml
[project]
name = "viz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bokeh>=3.8.1",
    "datashader>=0.18.2",
    "embedding-atlas>=0.13.0",
    "holoviews>=1.22.1",
    "matplotlib>=3.10.7",
    "modal>=1.2.4",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "s3fs>=2025.12.0",
    "sqlalchemy>=2.0.44",
    "umap-learn>=0.5.9.post2",
    "vercel>=0.3.5",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
]
